+++ scripts/install-services.sh
@@ -79,6 +79,7 @@ modules=(
     server/templates/messages/directory_scan_started.html server/templates/messages/directory_scan_completed.html
     server/templates/messages/directories_scanned.html server/templates/messages/batch_resumed.html
     server/templates/messages/identification_group_completed.html
+    app/TVEpisodeMapping.py app/prompts/DirectoryEpisodesTV.py
     app/Prompt.py
     app/BatchControlHandler.py app/BatchProcessor.py app/BatchRunner.py app/MessageHandler.py
     interface/BatchConfiguration.py interface/BatchControl.py entity/BatchRequest.py constants/DMessage.py